requirements.txt
faiss-cpu==1.7.4
langchain==0.0.178
pypdf2==3.0.1
streamlit==1.22.0
python-dotenv==1.0.0